requirements_minimal.txt
pathlib 
numpy 
matplotlib 
pillow
opencv-python
scikit_learn
seaborn